+++ R/pdRegression.R
@@ -16,7 +16,7 @@ pdRegression.default <- function(object, pred.var, pred.grid, train, progress,
     if (is.matrix(pred) || is.data.frame(pred)) {
       pred <- pred[, 1L, drop = TRUE]
     }
-    stats::setNames(mean(pred, na.rm = TRUE), "y")
+    stats::setNames(mean(pred, na.rm = TRUE), "yhat")
   }, .progress = progress, .parallel = parallel, .paropts = paropts)
 }
 
@@ -31,7 +31,7 @@ pdRegression.gbm <- function(object, pred.var, pred.grid, train, progress,
     log <- utils::capture.output(z <- mean(stats::predict(object,
                                                           newdata = temp, ...),
                                            na.rm = TRUE))
-    stats::setNames(z, "y")
+    stats::setNames(z, "yhat")
   }, .progress = progress, .parallel = parallel, .paropts = paropts)
 }
 
@@ -42,7 +42,8 @@ pdRegression.xgb.Booster <- function(object, pred.var, pred.grid, train,
   plyr::adply(pred.grid, .margins = 1, .fun = function(x) {
     temp <- train
     temp[pred.var] <- x
-    stats::setNames(mean(stats::predict(object, newdata = data.matrix(temp), ...),
-                  na.rm = TRUE), "y")
+    stats::setNames(mean(stats::predict(object,
+                                        newdata = data.matrix(temp), ...),
+                         na.rm = TRUE), "yhat")
   }, .progress = progress, .parallel = parallel, .paropts = paropts)
 }