+++ slowtests/slowtests.R
@@ -7,13 +7,77 @@
 # Setup
 ################################################################################
 
+# Make sure all packges are up to date
+# update.packages()
+
 # Load required packages
-library(caret)
 library(pdp)
 
 # Load data
 data(boston)  # regression
 data(pima)  # classification
+
+
+################################################################################
+# gbm package
+################################################################################
+
+# Load required package
+library(gbm)
+
+
+# Regression -------------------------------------------------------------------
+
+# Fit a gbm to the Boston housing data
+set.seed(101)
+boston.gbm <- gbm(cmedv ~ .,
+                  data = boston,
+                  distribution = "gaussian",
+                  n.trees = 10000,
+                  cv.folds = 10,
+                  interaction.depth = 3,
+                  shrinkage = 0.01,
+                  verbose = TRUE)
+best.iter <- gbm.perf(boston.gbm, method = "cv")
+
+# Compare plots for `lstat`
+par(mfrow = c(1, 2))
+plot(partial(boston.gbm, pred.var = "lstat", n.trees = best.iter), type = "l",
+     main = "pdp::partial")
+plot(boston.gbm, i.var = "lstat", n.trees = best.iter,
+     continuous.resolution = 51, main = "gbm::plot.gbm")
+
+
+# Classification ---------------------------------------------------------------
+
+# Fit a gbm to the Pima indians diabetes data
+set.seed(101)
+pima.gbm <- gbm(unclass(diabetes) - 1 ~ .,
+                data = pima,
+                distribution = "bernoulli",
+                n.trees = 5000,
+                interaction.depth = 3,
+                shrinkage = 0.001,
+                verbose = TRUE)
+best.iter <- gbm.perf(pima.gbm, method = "OOB")
+summary(pima.gbm, n.trees = best.iter, las = 1)
+
+# Compare plots for `glucose`
+par(mfrow = c(1, 2))
+plot(partial(pima.gbm, pred.var = "glucose", n.trees = best.iter), type = "l",
+     main = "pdp::partial")
+plot(pima.gbm, i.var = "glucose", n.trees = best.iter,
+     continuous.resolution = 51, main = "gbm::plot.gbm")
+
+# Compare plots for `glucose` and `age`
+p <- plot(pima.gbm, i.var = c("glucose", "mass"), n.trees = best.iter,
+          continuous.resolution = 51)
+grid.arrange(
+  partial(pima.gbm, pred.var = c("glucose", "mass"), plot = TRUE,
+          chull = TRUE, progress = "text", n.trees = best.iter),
+  p,
+  ncol = 2
+)
 
 
 ################################################################################
@@ -23,67 +87,39 @@ data(pima)  # classification
 # Load required package
 library(randomForest)
 
-# Regression
+
+# Regression -------------------------------------------------------------------
+
+# Fit a random forest to the Boston housing data
 set.seed(101)
-boston.randomForest <- randomForest(cmedv ~ ., data = boston, importance = TRUE)
+boston.randomForest <- randomForest(cmedv ~ .,
+                                    data = boston,
+                                    ntree = 500,
+                                    importance = FALSE)
 
 # Partial dependence of rm and lstat on cmedv
-pdp1 <- partial(boston.randomForest, pred.var = "rm", plot = TRUE, 
-                rug = TRUE, progress = "text")
-pdp2 <- partial(boston.randomForest, pred.var = "lstat", plot = TRUE, 
-                rug = TRUE, progress = "text")
-pdp3 <- partial(boston.randomForest, pred.var = "lon", plot = TRUE, 
-                rug = TRUE, progress = "text")
-pdp4 <- partial(boston.randomForest, pred.var = c("rm", "lstat"), plot = TRUE, 
-                rug = TRUE, chull = TRUE, progress = "text")
-multiplot(pdp1, pdp2, pdp3, pdp4, ncol = 2)
+grid.arrange(
+  partial(boston.randomForest, pred.var = "rm", plot = TRUE),
+  partial(boston.randomForest, pred.var = "lstat", plot = TRUE),
+  partial(boston.randomForest, pred.var = c("rm", "lstat"), plot = TRUE),
+  ncol = 3
+)
 
 
+# Classification ---------------------------------------------------------------
 
-ctrl <- trainControl(method = "repeatedcv", number = 10, repeats = 10, verboseIter = TRUE)
+# Fit a random forest to the Pima indians diabetes data
 set.seed(101)
-boston.mod <- train(x = data.matrix(subset(boston, select = -cmedv)),
-                    y = boston$cmedv,
-                    method = "svmRadial",
-                    metric = "Rsquared",
-                    trControl = ctrl,
-                    tuneLength = 10)
-plot(boston.mod)
-plot(varImp(boston.mod))
-partial(boston.mod, pred.var = "rm", plot = TRUE)
+pima.randomForest <- randomForest(diabetes ~ .,
+                                  data = pima,
+                                  na.action = na.omit,
+                                  ntree = 500,
+                                  importance = FALSE)
 
-topPredictors <- function(object, n = 1L, ...) {
-  UseMethod("topPredictors")
-}
-topPredictors.train <- function(object, n = 1L, ...) {
-  imp <- varImp(object)$importance
-  imp <- imp[order(imp$Overall, decreasing = TRUE), , drop = FALSE]
-  rownames(imp)[seq_len(n)]
-}
-topPredictors(boston.mod, 2)
-pdp.top <- boston.mod %>%
-  partial(topPredictors(boston.mod, n = 2), plot = TRUE, 
-          chull = TRUE, progress = "text") %>%
-  print()
-
-set.seed(101)
-mushroom.rf <- train(x = subset(mushroom, select = -Edibility),
-                     y = mushroom$Edibility,
-                     method = "rf",
-                     trControl = trainControl(method = "cv", number = 10, 
-                                              verboseIter = TRUE),
-                     tuneLength = 22)
-
-pd.odor <- partial(mushroom.rf, pred.var = "odor")
-p1 <- ggplot(varImp(mushroom.rf), color = "black") +
-  theme_light() +
-  xlab("") +
-  theme_bw()
-p2 <- ggplot(pd.odor, aes(x = odor, y = y)) +
-  geom_bar(stat = "identity") +
-  geom_hline(yintercept = 0) +
-  xlab("Odor") +
-  ylab("Log odds of being edible") +
-  coord_flip() +
-  theme_bw()
-multiplot(p1, p2, ncol = 2)
+# Partial dependence of rm and lstat on cmedv
+grid.arrange(
+  partial(pima.randomForest, pred.var = "glucose", plot = TRUE),
+  partial(pima.randomForest, pred.var = "mass", plot = TRUE),
+  partial(pima.randomForest, pred.var = c("glucose", "mass"), plot = TRUE),
+  ncol = 3
+)